requirements.txt
gymnasium==0.28.1
numpy==1.24.3
torch==2.0.1
